[ACA-4728] move reusable unit test config to testing module
